+++ R/imports.R
@@ -3,11 +3,10 @@ TRUE
 
 #' @importFrom purrr map map_df map_dbl map_lgl map_dfr
 #' @importFrom dplyr tibble as_tibble tribble
-#' @importFrom rlang UQ UQS sym syms ':='
 #' @importFrom forcats lvls_union fct_expand
 TRUE
 
-#' @importFrom stats setNames 
+#' @importFrom stats setNames
 #' @importFrom utils head tail
 #' @importFrom tools file_path_sans_ext
 TRUE